fix(ci): python-coverage-shards — split oversized test_validation_worktree.py to satisfy first-party file line limit
